requirements.txt
requests==2.31.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
email-validator==2.0.0
python-multipart==0.0.6
